+++ server/src/db/migration/data/migrate-data-json/RemoteSettings.ts
@@ -35,20 +35,8 @@ export type RemoteSettings = {
 }
 
 export const initialRemoteSettings: RemoteSettings = {
-  tumblrKeys: [
-    'BaQquvlxQeRhKRyViknF98vseIdcBEyDrzJBpHxvAiMPHCKR2l',
-    'G4iZd6FBiyDxHVUpNqtOTDu4woWzfp8WuH3tTrT3MC16GTmNzq',
-    'y5uUQJYTCp15Nj3P80cLmNFqwSr1mxIhm3C4PCsGAfReydkF9m',
-    'IZiOt6PYazf4g0sYWVfpfebMITRFWmtlKq2UKe6l0RsqKHPgui',
-    'ATtwOUlruyVl8bEiHTnYcRpByEAzov2LtLEWOfDLqhPRZFmT4X'
-  ],
-  tumblrSecrets: [
-    'XWVCo7t0GMGkOAd9wsxMMkKPhQbl3RqauGzQtnzAnmHCJ7WdSn',
-    'RmoWUh844NqVdw7btWI6EYldJ91KhwJyfCKPtAIcuVokFtRYgS',
-    'xiEV5sJISJAwegJHTTLWtxnmFUkowxgMk2gOq4mc20VNLM2TpJ',
-    'Iw3yKgjfMvrKPNCcqdUyRuxCkYWYyRlrMdFUojRHVkSSADOKCT',
-    'cMM7xqJV1roUudEdBiZeOqv3n1H0pzNnGY1iAbp3oo3c29MXGq'
-  ],
+  tumblrKeys: [],
+  tumblrSecrets: [],
 
   tumblrKey: '',
   tumblrSecret: '',